rerun on latest model and with logs

+++ __mbpp_tests_with_input_claude/task_848/solution_task_848.py
@@ -1,2 +1,2 @@
 def area_trapezium(a, b, h):
-    return (a + b) * h / 2
+    return int(((a + b) * h) / 2)
